requirements.txt
colorama==0.4.6
python-dotenv==1.0.0
openai>=1.0.0
pydantic>=2.0.0
sounddevice==0.4.6
numpy>=1.26.0
scipy>=1.12.0
soundfile==0.12.1
requests==2.31.0
elevenlabs==0.2.26
pynput==1.7.6 